### 2020-05-22 #2210 @whedon
[PRE REVIEW]: policytree: Policy learning via doubly robust empirical welfare maximization over trees
**Submitting author:** @erikcs (<a href="http://orcid.org/0000-0001-6093-1390">Erik Sverdrup</a>)
**Repository:** <a href="https://github.com/grf-labs/policytree" target ="_blank">https://github.com/grf-labs/policytree</a>
**Version:** v0.9.2
