### 2020-05-22 #2210 @erikcs
Thank you, it seems _AvianaGlobal_, _jjharden_, _nhejazi_  (causal inference) and _bblum9_ (statistics and C++) are good candidates.
