### 2020-05-22 #2210 @arfon
:wave: @bblum9, @nhejazi, @jjharden, @AvianaGlobal - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
