### 2020-05-22 08ad628
Address some inconsistencies in docstring notation
