### 2020-05-22 a39ab89
Update the docstring example to same as README
