### 2020-05-22 7c175f9
Update `multi_causal_forest` docstring (#19)
