+++ man/predict.multi_causal_forest.Rd
@@ -26,12 +26,17 @@ Computes estimates of \eqn{\tau_a(x)}
 }
 \examples{
 \donttest{
+# Train a multi causal forest.
 n <- 250
 p <- 10
+d <- 3
 X <- matrix(rnorm(n * p), n, p)
 W <- sample(c("A", "B", "C"), n, replace = TRUE)
 Y <- X[, 1] + X[, 2] * (W == "B") + X[, 3] * (W == "C") + runif(n)
-mcf <- multi_causal_forest(X = X, Y = Y, W = W)
-head(predict(mcf)$predictions)
+multi.forest <- multi_causal_forest(X = X, Y = Y, W = W)
+
+# Predict using the forest.
+multi.forest.pred <- predict(multi.forest)
+head(multi.forest.pred$predictions)
 }
 }
