### 2020-05-25 7b83cf9
Update vignette with mcf example
