### 2020-05-26 3e42a77
Simplify some headings in the intro vignette (#29)
