+++ vignettes/policytree.Rmd
@@ -7,6 +7,8 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+This document provides a short introduction to the `policytree` package, with examples from Zhou, Athey and Wager (2018), and Wager and Athey (2017). The last section addresses details, such as treatment estimates with one-vs-all grf, and the runtime of `policy_tree`.
+
 ```{r, echo = FALSE, message = FALSE}
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 set.seed(42)
@@ -17,7 +19,7 @@ library(policytree)
 library(grf)
 ```
 
-# Binary treatment effect estimation and policy learning
+# Ex. 1: Binary treatment effect estimation and policy learning
 ```{r}
 n <- 10000
 p <- 10
@@ -43,7 +45,7 @@ plot(X[, 1], X[, 2], col = pp)
 abline(0, -1, lwd = 4, col = 4)
 ```
 
-# Multi-action treatment effect estimation (Zhou, Athey and Wager, 2018)
+# Ex. 2: Multi-action treatment effect estimation
 
 The following example is from the 3-action DGP from section 6.4.1 in [Zhou, Athey and Wager (2018)](https://arxiv.org/abs/1810.04778)
 
@@ -96,7 +98,7 @@ leg <- sort(unique(pp))
 legend("topleft", legend = leg - 1, col = leg, pch = 10)
 ```
 
-# Efficient Policy Learning - Binary Treatment and Instrumental Variables (Wager and Athey, 2017)
+# Ex. 3: Efficient Policy Learning - Binary Treatment and Instrumental Variables
 
 The following example is from section 5.2 in [Wager and Athey (2017)](https://arxiv.org/abs/1702.02896).
 
