### 2020-06-05 #2232 @jjharden
I have completed a review of this paper and software. Overall, I believe it represents a useful contribution. Here are my comments for revision.

1. The documentation is somewhat sparse. I left a few items unchecked above which I think could be improved (statement of need, example, functionality). Also, there is a dependency on the DiagrammeR package that tripped up my initial run of the simulated data example from the paper.

2. I think the paper itself could better address "high-level functionality and [the] purpose of the software for a diverse, non-specialist audience."

I congratulate the authors on their contribution. Please let me know if you have further questions.
