### 2020-06-05 #2232 @erikcs
Thank you @jjharden.

1. Do you have suggestions for what you would see documented? Currently the introduction is in the online vignette: https://grf-labs.github.io/policytree/articles/policytree.html, with the rest being method docstrings.

    Not including the DiagrammeR is intentional to keep dependencies to a minimum. It is an optional add-on for the user who wishes to plot, you should have gotten an informative error message? ([this one](https://github.com/grf-labs/policytree/blob/master/R/policy_tree-plot.R#L89), it is done the same way in [grf](https://github.com/grf-labs/grf))
