### 2020-06-06 #2232 @jjharden
Sorry, my comments about documentation referred to the help files in R (e.g., by typing ?double_robust_scores). The online vignette looks complete.

Yes, that is the error message I got and I agree it was informative. I understand and appreciate wanting to minimize dependencies. It might be worth considering a footnote about the need for DiagrammeR in the paper. My sense is that most users would expect that code to run error-free. But it is a minor issue either way.
