### 2020-06-06 grf-labs/policytree#30 @nhejazi
[JOSS] Statement of Need
The JOSS review requires "Do the authors clearly state what problems the software is designed to solve and who the target audience is?" This is not currently addressed in the introductory section of the `README`. I'd recommend resolving this by simply copying a (possibly amended) bit from the JOSS paper into the `README` (an example of convenience, in my own packages I use sections like https://github.com/nhejazi/medshift#whats-medshift).
