### 2020-06-06 grf-labs/policytree#31 @nhejazi
[JOSS] Community Guidelines
The JOSS review requires "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support." The current contributions section (https://github.com/grf-labs/policytree#contributing) appears to be insufficient for this purpose. The currently included _issue template_ is a good start but there directions for contributing to the software could be improved. To resolve this, I would recommend

1. adding a section in the `README` clearly pointing to issues -- e.g., https://github.com/nhejazi/medshift#issues;
2. adding a document that describes the workflow for contributing (possibly including community expectations) -- e.g., https://github.com/nhejazi/medshift/blob/master/CONTRIBUTING.md;
3. updating the section in the `README` to clearly point to the document outlining contribution guidelines and community expectations -- e.g., https://github.com/nhejazi/medshift#contributions.

The above are examples of convenience (from my own work), but many alternative approaches are also perfectly acceptable.
