### 2020-06-06 #2232 @nhejazi
I just had some time to go through the package and paper throughout yesterday
and today. I agree with @jjharden that the package seems to provide an good and
very useful contribution. For a few major comments about the package, I opened
up issues https://github.com/grf-labs/policytree/issues/30 and
https://github.com/grf-labs/policytree/issues/31. Once these are addressed, I'll
be able to check off the few remaining items on the checklist above. Some minor
comments on the package and paper are included below.

Minor comments on the package (these need not be addressed for acceptance):
* The package documentation generally looks good, though there are minor aspects
  of it that could be improved, such as adding short running examples for
  exported functions/methods (e.g., `double_robust_scores.causal_forest` and
  `conditional_means.causal_forest`).
* It would be nice to add a citations section both to the `README` and as a
  plain text file in the `inst` directory (e.g.,
  https://github.com/nhejazi/medshift/blob/master/inst/CITATION), so that
  citations can be generated from R via `citation(policytree)`.
* The `README.md` is generally well written. It includes appropriate R examples
  and references, both good things. Since the file includes example code (and
  the output of this code), I would recommend converting the `README.md` file
  to an R Markdown file `README.Rmd` that can be used to auto-generate the
  `README.md`. This will be useful to ensure that the example code remains up
  to date as the package evolves over time. For example, consider
  https://github.com/nhejazi/medshift/blob/master/README.Rmd

Minor comments on the paper:
* The first sentence of the second paragraph should read: "There has recently
  been a considerable amount __[of]__ work..." (missing preposition added).
* The R language is mentioned but not cited. For this purpose, I generally use
  a BibTeX entry like the following https://scholar.googleusercontent.com/scholar.bib?q=info:vROS1kQC7b8J:scholar.google.com/&output=citation&scisdr=CgXI3MF1EK2KvDjnRqI:AAGBfm0AAAAAXtriXqLBfUBBIBiwusVRnSFBi3CX4VsU&scisig=AAGBfm0AAAAAXtriXmJOD9c7swu-rkbMdWU5sgDW3caO&scisf=4&ct=citation&cd=-1&hl=en
  (but update the years). Feel free to use another appropriate entry too.
* A good overview of methods for learning optimal rules is provided. It could
  be improved by including other recent doubly robust and data adaptive
  developments in similar vein, e.g., https://pubmed.ncbi.nlm.nih.gov/27227726/
  an https://pubmed.ncbi.nlm.nih.gov/26236571/
* Since you asked @erikcs, an R package implementing the above approaches is
  available https://github.com/tlverse/tmle3mopttx, documented in long-form at
  https://tlverse.org/tlverse-handbook/optimal-individualized-treatment-regimes.html
