### 2020-06-07 a59ffb9
Add more details to Contributing
