### 2020-06-07 c3bb981
Add a better package description to README
