### 2020-06-07 4056668
Update scores docstring with causal forest example
