+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 The problem of learning treatment assignment policies from randomized or observational data arises in many fields. For example, in personalized medicine, we seek to map patient observables (like age, gender, heart pressure, etc.) to a treatment choice using a data-driven rule.
 
-There has recently been a considerable amount work on statistical methodology for policy learning, including @manski2004statistical, @zhao2012estimating, @swaminathan2015batch, @kitagawa2018should, @mbakop2016model, @athey2017efficient, @kallus2018confounding and @zhou2018offline. In particular, @kitagawa2018should show that if we only consider policies $\pi$ restricted to a class $\Pi$ with finite VC dimension and have access to data from a randomized trial with $n$ samples, then an empirical welfare maximization algorithm achieves regret that scales as $\sqrt{\operatorname{VC}(\Pi) / n}$. @athey2017efficient extend this result to observational studies via doubly robust scoring, and @zhou2018offline further consider the case with multiple treatment choices (in particular, the regret will depend on the tree depth, feature space, and number of actions).
+There has recently been a considerable amount of work on statistical methodology for policy learning, including @manski2004statistical, @zhao2012estimating, @swaminathan2015batch, @kitagawa2018should, @mbakop2016model, @athey2017efficient, @kallus2018confounding and @zhou2018offline. In particular, @kitagawa2018should show that if we only consider policies $\pi$ restricted to a class $\Pi$ with finite VC dimension and have access to data from a randomized trial with $n$ samples, then an empirical welfare maximization algorithm achieves regret that scales as $\sqrt{\operatorname{VC}(\Pi) / n}$. @athey2017efficient extend this result to observational studies via doubly robust scoring, and @zhou2018offline further consider the case with multiple treatment choices (in particular, the regret will depend on the tree depth, feature space, and number of actions).
 
 The package ``policytree`` for ``R`` implements the multi-action doubly robust approach of @zhou2018offline in the case where we want to learn policies $\pi$ that belong to the class $\Pi$ of depth-$k$ decision trees. In order to use ``policytree``, the user starts by specifying a set of doubly robust scores for policy evaluation; the software then carries out globally optimal weighted search over decision trees.
 
