+++ paper/paper.bib
@@ -116,3 +116,12 @@
   year={2015},
   publisher={JMLR. org}
 }
+
+@manual{R,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2020},
+  url = {https://www.R-project.org/}
+}
