### 2020-06-07 a52ca3a
Update the R cite type
