+++ paper/paper.bib
@@ -125,3 +125,25 @@
   year = {2020},
   url = {https://www.R-project.org/}
 }
+
+@article{luedtke2016super,
+  title={Super-learning of an optimal dynamic treatment rule},
+  author={Luedtke, Alexander R and van der Laan, Mark J},
+  journal={The international journal of biostatistics},
+  volume={12},
+  number={1},
+  pages={305--332},
+  year={2016},
+  publisher={De Gruyter}
+}
+
+@article{van2015targeted,
+  title={Targeted learning of the mean outcome under an optimal dynamic treatment rule},
+  author={van der Laan, Mark J and Luedtke, Alexander R},
+  journal={Journal of causal inference},
+  volume={3},
+  number={1},
+  pages={61--95},
+  year={2015},
+  publisher={De Gruyter}
+}
