+++ paper/paper.bib
@@ -147,3 +147,11 @@
   year={2015},
   publisher={De Gruyter}
 }
+
+@manual{tmle3,
+  title = {tmle3mopttx: Targeted Maximum Likelihood Estimation of the Mean under Optimal Individualized Treatment},
+  author = {Ivana Malenica and Jeremy Coyle and Mark {van der Laan}},
+  year = {2020},
+  note = {R package version 0.1.0},
+  url = {https://tlverse.org/tmle3mopttx},
+}
