+++ paper/paper.bib
@@ -128,7 +128,7 @@
 
 @article{luedtke2016super,
   title={Super-learning of an optimal dynamic treatment rule},
-  author={Luedtke, Alexander R and van der Laan, Mark J},
+  author={Luedtke, Alexander R and {van der Laan}, Mark},
   journal={The international journal of biostatistics},
   volume={12},
   number={1},
@@ -139,7 +139,7 @@
 
 @article{van2015targeted,
   title={Targeted learning of the mean outcome under an optimal dynamic treatment rule},
-  author={van der Laan, Mark J and Luedtke, Alexander R},
+  author={{van der Laan}, Mark and Luedtke, Alexander R},
   journal={Journal of causal inference},
   volume={3},
   number={1},
