### 2020-06-07 #2232 @erikcs
Thank you @nhejazi, I have opened three PRs in https://github.com/grf-labs/policytree/pulls.

The paper has also been updated with more references.
