### 2020-06-07 #2232 @erikcs
> ☝️ @erikcs - commands to Whedon have to be standalone in their own comment (or at the very start of a comment).

Got it, thanks.
