### 2020-06-07 f74b204
Add a note that DiagrammeR is needed for tree plots
