+++ paper/paper.md
@@ -37,7 +37,7 @@ The package ``policytree`` for ``R`` [@R] implements the multi-action doubly rob
 
 It is well known that finding an optimal tree of arbitrary depth is NP-hard. However, if we restrict our attention to trees of depth $k$, then the problem can be solved in polynomial time. Here, we implement the global optimization via an exhaustive (unconstrained) tree search that runs in $O(P^{k} N^{k} (\log N + D) + PN\log N)$ time, where $N$ is the number of individuals, $P$ the number of characteristics observed for each individual and $D$ is the number of available treatment choices (see details below). If an individual’s characteristics only takes on a few discrete values, the runtime can be reduced by a factor of $N^k$. Additionally, an optional approximation parameter lets the user control how many splits to consider.
 
-![A depth 2 tree fit on data from the National Job Training Partnership Act Study [@bloom1997benefits]. The reward matrix contains two outcomes: not assigning treatment (action 1), and assigning treatment, a job training program (action 2). The covariate matrix contains two variables: a candidate's previous annual earnings in \$1,000 and years of education.](example.png)
+![A depth 2 tree fit on data from the National Job Training Partnership Act Study [@bloom1997benefits]. The reward matrix contains two outcomes: not assigning treatment (action 1), and assigning treatment, a job training program (action 2). The covariate matrix contains two variables: a candidate's previous annual earnings in \$1,000 and years of education. Note: the optional package ```DiagrammeR``` is needed to plot trees.](example.png)
 
 Our package is integrated with the ``R`` package ``grf`` of @athey2019generalized, allowing for a simple workflow that uses random forests to estimate the nuisance components required to automatically form the doubly robust scores. We also generalize the ``causal_forest`` function from ``grf`` to multiple treatment effects with a one vs all encoding described in @zhou2018offline. The following simulation example illustrates this workflow in a setting with $D = 3$ actions; here, we write covariates with `X`, outcomes as `Y`, and actions as `W`. Figure 1 shows a tree similarly grown on a dataset considered by @kitagawa2018should.
 
