### 2020-06-07 #2232 @erikcs
@jjharden I added a note under Figure 1 that `DiagrammeR` is needed for tree plots.
