### 2020-06-11 e4d0e48
Reorder the van der Laan cites in order of first paper appearance date
