### 2020-06-18 67b3be3
Move the contribution guidelines into a separate DEVELOPING.md file
