### 2020-06-19 #2232 @erikcs
Just checking in and summarizing the changes so far:

1. Paper: under Figure 1 we add a note that DiagrammeR is needed for tree plots (we do not include it as a dependency since it has a ton of dependencies).

2. Paper: we fix the typo pointed out

3. Paper: we add add two articles pointed out by @nhejazi to the citation list, cite the use of R, and cite another useful package from the tlverse.

4. Code: we update the dosctring for `double_robust_scores` and `conditional_means` with one more example (https://github.com/grf-labs/policytree/pull/34)

5. Code: We add a more informative package description on the GitHub landing page (https://github.com/grf-labs/policytree/pull/32)

6. Code: We clarify the community guidelines by adding a separate DEVELOPING.md (this reuses some info from our parent package, grf), and add more descriptive links to issues and the online documentation (https://github.com/grf-labs/policytree/pull/33).
