### 2020-06-19 #2232 @jjharden
The paper looks ready to go to me. I have no more suggested changes.
