### 2020-06-21 #2232 @arfon
Thanks @nhejazi and @jjharden.

@erikcs - could you check those potentially missing DOIs for the references and also give the paper one final read to check everything looks OK?
