### 2020-06-21 2ab669a
Update scores/means documentation with causal forest example (#34)
