### 2020-06-21 70f5485
Clarify the community guidelines in README (#33)
