+++ paper/paper.bib
@@ -13,7 +13,9 @@
   number={2},
   pages={591--616},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.3982/ECTA13288},
+  url = {https://doi.org/10.3982/ECTA13288}
 }
 
 @article{zhou2018offline,
@@ -48,7 +50,9 @@
   volume={32},
   number={3},
   year={1997},
-  publisher={University of Wisconsin Press}
+  publisher={University of Wisconsin Press},
+  doi = {10.2307/146183},
+  url = {https://doi.org/10.2307/146183}
 }
 
 @inproceedings{kallus2018confounding,
@@ -92,7 +96,9 @@
   number={4},
   pages={1221--1246},
   year={2004},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1111/j.1468-0262.2004.00530.x},
+  url = {https://www.jstor.org/stable/3598783}
 }
 
 @article{zhao2012estimating,
@@ -134,7 +140,9 @@
   number={1},
   pages={305--332},
   year={2016},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi = {10.1515/ijb-2015-0052},
+  url = {https://doi.org/10.1515/ijb-2015-0052}
 }
 
 @article{van2015targeted,
