### 2020-06-22 #2232 @erikcs
> @whedon generate pdf from branch oss-paper
>
> Thanks @nhejazi and @jjharden.
>
> @erikcs - could you check those potentially missing DOIs for the references and also give the paper one final read to check everything looks OK?

Thank you @arfon, I have updated the .bib file with DOIs from the publisher's page for the listed items.
