+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: policytree
 Title: Policy Learning via Doubly Robust Empirical Welfare Maximization over Trees
-Version: 0.9.2
+Version: 1.0
 Authors@R: c(
     person("Zhengyuan", "Zhou", role = "aut"),
     person("Susan", "Athey", role = "aut"),
