+++ releases/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to policytree will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0] - 2020-06-22
+
+Release 1.0 has been reviewed by the Journal of Open Source Software. See https://github.com/openjournals/joss-reviews/issues/2232 for improvements.
+
+### Fixed
+- Fix treatment vector categories bug. [#22](https://github.com/grf-labs/policytree/pull/22)
+
 ## [0.9.2] - 2020-03-21
 
 ### Fixed
