### 2020-06-22 #2232 @erikcs
@arfon A version 1.0 version of the [package](https://github.com/grf-labs/policytree/pull/35) with the enhancements arising from this review (and some earlier fixes) is published on [CRAN](https://cran.r-project.org/web/packages/policytree/index.html) now.

The Zenodo archive is here: https://zenodo.org/record/3903401
