### 2020-06-22 #2232 @arfon
@erikcs - thanks for this. One small request, please use a Zenodo-issued DOI for your deposit with them. You've currently got the _JOSS_ DOI listed as the DOI but this won't work:
