### 2020-06-22 #2232 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
